gpu/recompiler: split rdna2_to_spirv.cpp per shader stage
A pure move: no line edited, only relocated. split_file.py proves it -- the five outputs
read back from disk rebuild the original 206,055 bytes exactly.

  rdna2_to_spirv.cpp             582   entry points and dispatch
  rdna2_recompile_vertex.cpp     940   vertex, incl. NGG passthrough analysis
  rdna2_recompile_compute.cpp    968   compute, LDS fmin/fmax sync, coverage
  rdna2_recompile_fragment.cpp   631   fragment, incl. interpolation layout
  spirv_wave_width.cpp           695   wave-width independence, a SPIR-V-level pass
  rdna2_recompile_shared.hpp   1,196   shared analyses (previous commit)

The stages were already separate in the reference graph -- the seam is where the tool
found 1% of references crossing, not where the names suggested. spirv_wave_width is the
one that is not a stage: it rewrites emitted SPIR-V rather than decoding RDNA2, which is
why it reads as its own group and gets its own file.

Verified: builds clean; ctest --no-tests=error -j4 reports 457/457, exit 0.

This split failed on the first attempt and the failure is why #3524 and #3525 exist. It
passed every check -- tiling, partition, byte-identical reconstruction, and the cross-part
reference check -- and then would not compile, because that check tested LINKAGE where the
question is whether a DECLARATION is in scope. `safe_execz_branches` has external linkage
and no header declares it, so three of these five files called a function that links and
does not compile. The map records `declared_in_header` now.
